--- a/packages/flutter/lib/src/widgets/framework.dart
+++ b/packages/flutter/lib/src/widgets/framework.dart
@@ -14,6 +14,7 @@ import 'debug.dart';
 import 'focus_manager.dart';
 import 'inherited_model.dart';
 import 'notification_listener.dart';
+import 'widget_inspector.dart';
 
 export 'package:flutter/foundation.dart' show
   factory,
@@ -2640,7 +2641,8 @@ class BuildOwner {
           }
           return true;
         }());
-        if (!kReleaseMode && debugProfileBuildsEnabled) {
+        final bool isTimelineTracked = !kReleaseMode && _isProfileBuildsEnabledFor(element.widget);
+        if (isTimelineTracked) {
           Map<String, String> debugTimelineArguments = timelineArgumentsIndicatingLandmarkEvent;
           assert(() {
             debugTimelineArguments = element.widget.toDiagnosticsNode().toTimelineArguments();
@@ -2668,7 +2670,7 @@ class BuildOwner {
             ],
           );
         }
-        if (!kReleaseMode && debugProfileBuildsEnabled)
+        if (isTimelineTracked)
           Timeline.finishSync();
         index += 1;
         if (dirtyCount < _dirtyElements.length || _dirtyElementsNeedsResorting!) {
@@ -3076,6 +3078,13 @@ class _NotificationNode {
     }
     parent?.dispatchNotification(notification);
   }
+}
+
+bool _isProfileBuildsEnabledFor(Widget widget) {
+  // ignore: avoid_bool_literals_in_conditional_expressions
+  return debugProfileBuildsEnabled && debugProfileBuildsEnabledUserWidgets
+      ? debugIsWidgetLocalCreation(widget)
+      : true;
 }
 
 /// An instantiation of a [Widget] at a particular location in the tree.
@@ -3503,7 +3512,8 @@ abstract class Element extends DiagnosticableTree implements BuildContext {
       } else if (hasSameSuperclass && Widget.canUpdate(child.widget, newWidget)) {
         if (child.slot != newSlot)
           updateSlotForChild(child, newSlot);
-        if (!kReleaseMode && debugProfileBuildsEnabled) {
+        final bool isTimelineTracked = !kReleaseMode && _isProfileBuildsEnabledFor(newWidget);
+        if (isTimelineTracked) {
           Map<String, String> debugTimelineArguments = timelineArgumentsIndicatingLandmarkEvent;
           assert(() {
             debugTimelineArguments = newWidget.toDiagnosticsNode().toTimelineArguments();
@@ -3515,7 +3525,7 @@ abstract class Element extends DiagnosticableTree implements BuildContext {
           );
         }
         child.update(newWidget);
-        if (!kReleaseMode && debugProfileBuildsEnabled)
+        if (isTimelineTracked)
           Timeline.finishSync();
         assert(child.widget == newWidget);
         assert(() {
@@ -3765,7 +3775,8 @@ abstract class Element extends DiagnosticableTree implements BuildContext {
   Element inflateWidget(Widget newWidget, Object? newSlot) {
     assert(newWidget != null);
 
-    if (!kReleaseMode && debugProfileBuildsEnabled) {
+    final bool isTimelineTracked = !kReleaseMode && _isProfileBuildsEnabledFor(newWidget);
+    if (isTimelineTracked) {
       Map<String, String> debugTimelineArguments = timelineArgumentsIndicatingLandmarkEvent;
       assert(() {
         debugTimelineArguments = newWidget.toDiagnosticsNode().toTimelineArguments();
@@ -3800,7 +3811,7 @@ abstract class Element extends DiagnosticableTree implements BuildContext {
     newChild.mount(this, newSlot);
     assert(newChild._lifecycleState == _ElementLifecycle.active);
 
-    if (!kReleaseMode && debugProfileBuildsEnabled)
+    if (isTimelineTracked)
       Timeline.finishSync();
 
     return newChild;