Turning on widget creation tracking for profile builds allows us to do performance evaluation techniques that wouldn't scale across all widgets.  In this case we are using it to track build time for widgets on the CPU but it will also allow us to start tracking raster performance of widgets (cc @iskakaushik)

Added overhead (profile builds):
 * memory
   * ~50 bytes per widget for the construction location (I don't believe we are taking advantage of `const` to reduce this)
   * Memoization of `_isLocalCreationLocation`, ~5 bytes per source file used in the app that constructs widgets.
 * cpu
   *  An extra timeline event for each user created widget's `build`
   * Extra parameter for each widget constructor call
   * String comparisons against construction file uri (memoized with HashMap) for build of each widget

fixes https://github.com/flutter/flutter/issues/100928

## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I read and followed the [Flutter Style Guide], including [Features we expect every widget to implement].
- [x] I signed the [CLA].
- [x] I listed at least one issue that this PR fixes in the description above.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] I added new tests to check the change I am making, or this PR is [test-exempt].
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-new channel on [Discord].

<!-- Links -->
[Contributor Guide]: https://github.com/flutter/flutter/wiki/Tree-hygiene#overview
[Tree Hygiene]: https://github.com/flutter/flutter/wiki/Tree-hygiene
[test-exempt]: https://github.com/flutter/flutter/wiki/Tree-hygiene#tests
[Flutter Style Guide]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo
[Features we expect every widget to implement]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo#features-we-expect-every-widget-to-implement
[CLA]: https://cla.developers.google.com/
[flutter/tests]: https://github.com/flutter/tests
[breaking change policy]: https://github.com/flutter/flutter/wiki/Tree-hygiene#handling-breaking-changes
[Discord]: https://github.com/flutter/flutter/wiki/Chat