@gaaclarke 2022-03-29
I measured the BUILD times of the Flutter Gallery app to make sure we aren't introducing a lot of overhead by turning on track-widget-creation on in profile builds.  The Flutter Gallery has 3 BUILD's before the first LAYOUT.  Here are the results:

| type | 1st BUILD | 2nd BUILD | 3rd BUILD |
|-|-|-|-|
| w/ track-widget-creation | 6.3 ms | 1.8 ms | 2.1 ms |
| w/o track-widget-creation | 6ms | 1.8 ms | 2.8 ms |

This was tested on an iOS device.  At first blush there doesn't appear to be a significant change in BUILD times as a result of tracking widget creation.