@gaaclarke 2022-03-29
I ran a performance test with `debugProfileBuildsEnabledUserWidgets` turn off an on as it is currently written:
|debugProfileBuildsEnabledUserWidgets|1st BUILD|2nd BUILD|3rd BUILD|
|-|-|-|-|
|off|6.3 ms|1.8 ms|2.1 ms|
|on|6.5 ms|1.5 ms|2.7 ms|