@gaaclarke 2022-03-29
@Hixie I chose you to review this since you've done work on the Timeline events and we talked about this feature in the past.