@gaaclarke 2022-03-31
@dnfield some of the comments were similar so I'll respond to them here.

> Can we not default this to true? \
\
Alternatively, if we do default this true, please provide some numbers on how this impacts internal benchmarks we are tracking. The idea is we should know exactly how much this regresses/adds noise to those benchmarks.

> Please check this patch on the internal mobile harness testing on a low end device. Flutter Gallery even on a Moto G4 is not representative, particularly for things that will do extra work on every frame, compared to the internal low end Android device for customer: money.

> Please do not land this until we've answered this question. Ideally, we'd know how much overhead just this line adds, compared to the rest of this patch, as measured on a low end Android device with a large application like customer: money.

The patch was originally written with the flag set to `false`.  The performance of the the patch was encouraging and lead me to believe the utility outweighs any cost.  Here's some more datapoints that support that decision:
1) We already create a comparable number of Timeline events in the LAYOUT phase.
1) We know from the Timeline benchmarks that Timeline events without parameters are much faster since they aren't doing JSON encoding.  So we have gained cycles there.
1) Any added overhead only effects Profile builds.  If someone is running a Profile build, presumably they want information about the performance of the app, so this is information they want and shouldn't be hidden behind a flag (assuming the cost is reasonable).
1) We know from the Timeline benchmark that it's actually much slower on iOS devices, so testing on iOS is more informative.

Any change to the Framework can adversely affect performance, the bar for landing Framework PR's isn't proving it doesn't have any effect.  The calculus is (user value / performance overhead).  Because this change is fundamental to the Framework I've done due diligence by explaining all performance differences and measuring them locally.

If you have a way I can test customer:money preemptively I'm happy to to do that.  I'm assuming I can apply the patch to a workspace and run a script to get the results from a remote device?  Give me details on what you'd like to test precisely and what results would be acceptable.  If the bar is zero cost to profile builds, that might not be the case, that depends on the savings we get from not doing json encoding in the Timeline event.  If the bar is less than a millisecond for the meanest reasonable scene we can throw at it, I think thats possible.

Maybe a better check would be a microbenchmark with N user widgets?