@gaaclarke 2022-03-31
@Hixie @dnfield it turns out we already have a good microbenchmark on build times, `flutter/dev/benchmarks/microbenchmarks/lib/stocks/build_bench.dart`

That gives better results than the one I listed above:
|test| result|
|-|-|
| before change | 831.9 µs |
| after change (flag turned off) | 826.1 µs |
| after change (flag turned on) | 969.4 µs |

That shows that the addition of tracking the widget creation and removing the diagnostic nodes in profile builds is a net positive.  The diagnostic nodes had no meaningful data and were just causing JSON encoding to happen.

The net change with the flag on an iOS device (where it will be higher) was 137.5 µs.  Even if that value is off by an order of magnitude it is an acceptable cost.

If you don't think that test is representative let me know and we can tweak it.