@dnfield 2022-03-31
> Any change to the Framework can adversely affect performance, the bar for landing Framework PR's isn't proving it doesn't have any effect. 

This particular change is touching an are of the framework that is very performance sensitive, so the bar has to be higher for it. And because it has a high likelyhood of impacting a high priority customer's benchmarks, we really should do more DD up front before having a fire drill after a google3 roll that contains 100 commits.

> The net change with the flag on an iOS device (where it will be higher) was 137.5 µs. Even if that value is off by an order of magnitude it is an acceptable cost.

I do not agree. That represents a >10% regression on a microbenchmark. That seems unacceptably high to me to opt this in. Especially since this feature is really providing performance debugging information that is not essential to running an application or otherwise doing performance testing. I am happy to see that there's a slight improvement with the flag turned off, and I would be more willing to approve this PR if it were defaulted to off even if we don't do the internal testing - since I think it's pretty reasonable to assume it will have a net neutral impact on internal benchmarks in that case.