@gaaclarke 2022-03-31
> That represents a >10% regression on a microbenchmark.

Right, but I would like to point out that it isn't really a scalar on the build time.  If a widget took 200% longer to build the flag would still cost 137.5 µs.  The important measurement is the absolute added overhead given the most complex scene with the worst setup.  I think we can say that that will be likely be below 1ms.

> I would be more willing to approve this PR if it were defaulted to off even if we don't do the internal testing - since I think it's pretty reasonable to assume it will have a net neutral impact on internal benchmarks in that case.

Okay, I'll start working on cleaning everything up since it sounds like we can go ahead one way or another.

> (And I do really appreciate you getting these numbers, it's a big help)

No problem.