@gaaclarke 2022-03-31
Also in a different PR I'm tweaking the microbenchmarks to also report the probability of the results.  I'm not sure if that will inform the results but it will be nice to see.