@gaaclarke 2022-04-01
Okay, all the cleanup is done.  I'll try a few different things to see if we can get the performance better, otherwise we can turn this off by default.