@gaaclarke 2022-04-01
I tweaked `_isLocalCreationLocation` to just randomly return true in order to isolate the cost of that function versus the Timeline cost.  The results showed that the 137 µs overhead is mostly the cost of doing Timeline events, so there is no way we can tweak that logic to get the performance we want, we have to make Timeline events faster.

I also compared the performance of tracing all widgets versus user-created widgets on the stocks microbenchmark to see what kind of savings we get when using this flag.
|test|result|
|-|-|
|all widgets| 3308.9 µs |
|user-created widgets| 972.0 µs |

I'll turn this off by default and I'll look into integrating with the work @iskakaushik is doing to turn this feature on temporarily to give performance information on demand.